强化（没测试）

+++ Assets/Scripe/Utils/InventoryHalper.cs
@@ -5,7 +5,7 @@ using System;
 public class InventoryHalper 
 {
     List<PlayerBackpackBean> mList = new List<PlayerBackpackBean>();
-    private Dictionary<long, PlayerBackpackBean> mRoleUseList = new Dictionary<long, PlayerBackpackBean>();
+    private List< PlayerBackpackBean> mRoleUseList = new List<PlayerBackpackBean>();
     private Dictionary<long, long> mDropDeviceUsed = new Dictionary<long, long>();
     private List<long> mHaveBookId = new List<long>();
 
@@ -24,8 +24,8 @@ public class InventoryHalper
         List<PlayerBackpackBean> list = SQLHelper.getIntance().getPlayUserZhuangbei();
         if (list != null && list.Count > 0) {
             foreach (PlayerBackpackBean b in list) {
-                AccouterJsonBean acc = BackpackManager.getIntance().getAccouterInfoById(b.goodId);
-                mRoleUseList.Add(acc.type, b);
+       //         AccouterJsonBean acc = BackpackManager.getIntance().getAccouterInfoById(b.goodId);
+                mRoleUseList.Add(b);
             }
         }
     }
@@ -117,15 +117,35 @@ public class InventoryHalper
                 newBean.sortID = jb.sortID;
                 newBean.count = count;
                 newBean.tabId = jb.tabid;
-                newBean.attributeList = new List<PlayerAttributeBean>();重写
-                foreach (AttributeBean be in jb.getAttributeList())
+                newBean.attributeList = new List<PlayerAttributeBean>();
+                foreach (EquipKeyAndValue be in jb.getAttributeList())
                 {
-                    PlayerAttributeBean p = new PlayerAttributeBean();
-                    p.type = be.type;
-                    p.value = be.getCurrentValue();
-                    newBean.attributeList.Add(p);
+                    PlayerAttributeBean p1 = new PlayerAttributeBean();
+                    p1.type = be.key;
+                    p1.value = be.value;
+                    newBean.attributeList.Add(p1);
+                    Debug.Log("newBean.attributeList.Add : type=" + p1.type + " value=" + p1.value);
                 }
+                PlayerAttributeBean p2 = new PlayerAttributeBean();
+                p2.type = 10001;
+                p2.value = 0;
+                newBean.attributeList.Add(p2);
+                Debug.Log("newBean.attributeList.Add : type=" + p2.type + " value=" + p2.value);
+                long affix = jb.getAffix();
+                AffixJsonBean bean2 = JsonUtils.getIntance().getAffixInfoById(affix);
+                long value2 = bean2.getAffixValue(jb.quality);
+                PlayerAttributeBean p3 = new PlayerAttributeBean();
+                p3.type = 10002;
+                p3.value = affix;
+                newBean.attributeList.Add(p3);
+                Debug.Log("newBean.attributeList.Add : type=" + p3.type + " value=" + p3.value);
+                PlayerAttributeBean p4 = new PlayerAttributeBean();
+                p4.type = affix;
+                p4.value = value2;
+                newBean.attributeList.Add(p4);
+                Debug.Log("newBean.attributeList.Add : type=" + p4.type + " value=" + p4.value);
                 mList.Add(newBean);
+
             }
             else if (id > TABID_3_START_ID) {
                 CardJsonBean cj = BackpackManager.getIntance().getCardInfoById(id);
@@ -193,32 +213,43 @@ public class InventoryHalper
     }
 
 
-    public Dictionary<long, PlayerBackpackBean> getRoleUseList() {
+    public List< PlayerBackpackBean> getRoleUseList() {
         return mRoleUseList;
     }
     public List<long> getHaveBookId() {
         return mHaveBookId;
     }
-    public void use(PlayerBackpackBean bean, long count)
+    public void updateZhuangbei(PlayerBackpackBean bean, long level)
     {
-        PlayerBackpackBean beanOld = null;
-        AccouterJsonBean acc = BackpackManager.getIntance().getAccouterInfoById(bean.goodId);
-
-        if (mRoleUseList.ContainsKey(acc.type))
-        {
-            beanOld = mRoleUseList[acc.type];
-            deleteRoleUse(acc.type,bean);
-            
-            addRoleUse(acc.type, bean);
+        level = level + 1;
+       string extan =  SQLHelper.getIntance().getGoodExtra(bean);
+        AccouterJsonBean ac = JsonUtils.getIntance().getAccouterInfoById(bean.goodId);
+        foreach (PlayerAttributeBean p in bean.attributeList){
+            if (p.type >= 100 && p.type <= 999 && ac.getStrengThen().ContainsKey(p.type))
+            {
+                List<EquipKeyAndValue> ekv = ac.getStrengThen()[p.type];
+                for (int i = 0; i < ac.getAttributeList().Count; i++)
+                {
+                    if (ac.getAttributeList()[i].key == p.type)
+                    {
+                        p.value = ac.getAttributeList()[i].value;
+                        break;
+                    }
+                }
+                long value = ac.getStrengthenByLevel(p.type,level);
+                p.value = value + p.value;
+            }
+            else if(p.type == 10001)
+            {
+                p.value = level;
+            }
         }
-        else
-        {
-            addRoleUse(acc.type, bean);            
-        }
-        deleteIventory(bean);
-        if (beanOld != null) {
-            addIventory(bean);
-        }
+        SQLHelper.getIntance().updateZHUANGBEI(bean, extan);
+       
+    }
+    public bool use(PlayerBackpackBean bean, long count)
+    {        
+        return addRoleUse(bean);
     }
     private bool isAddSuccess = true;
     public long useBook(PlayerBackpackBean bean,long count) {
@@ -260,21 +291,32 @@ public class InventoryHalper
     }
 
     public void unUse(PlayerBackpackBean bean, long count) {
-        AccouterJsonBean acc = BackpackManager.getIntance().getAccouterInfoById(bean.goodId);
-        mRoleUseList.Remove(acc.type);
-        SQLHelper.getIntance().deleteZuangbei(bean.goodId);
-        addIventory(bean);
+        if (mRoleUseList.Remove(bean))
+        {
+            SQLHelper.getIntance().deleteZuangbei(bean.goodId);
+            addIventory(bean);
+        }
     }
 
-    private void deleteRoleUse(long type, PlayerBackpackBean bean)
+    private bool deleteRoleUse( PlayerBackpackBean bean)
     {
-        mRoleUseList.Remove(type);
-        SQLHelper.getIntance().deleteZuangbei(bean.goodId);
+        if (mRoleUseList.Remove(bean)) {
+            SQLHelper.getIntance().deleteZuangbei(bean.goodId);
+            
+            return true;
+        }
+        return false;
+        
+        
     }
-    private void addRoleUse(long type,PlayerBackpackBean bean)
+    private bool addRoleUse(PlayerBackpackBean bean)
     {
-        mRoleUseList.Add(type,bean);
-        SQLHelper.getIntance().addZHUANGBEI(bean);
+        if (mRoleUseList.Count < 6) {
+            mRoleUseList.Add( bean);
+            deleteIventory(bean);
+            SQLHelper.getIntance().addZHUANGBEI(bean);
+        }
+        return false;
     }
 
 
@@ -282,7 +324,7 @@ public class InventoryHalper
     {
         Debug.Log("mList size = " + mList.Count);
         mList.Remove(bean);
-        SQLHelper.getIntance().deleteGood(bean.goodId);
+        SQLHelper.getIntance().deleteGood(bean);
         Debug.Log("mList size = " + mList.Count);
     }
     private void addIventory(PlayerBackpackBean bean)
@@ -306,7 +348,7 @@ public class InventoryHalper
         if (bean.count == 0)
         {
             mList.Remove(bean);
-            SQLHelper.getIntance().deleteGood(bean.goodId);
+            SQLHelper.getIntance().deleteGood(bean);
         }
         else {
             SQLHelper.getIntance().ChangeGood(bean);
